rm unneeded bool from unit test, fix test file name Co-authored-by: Caden Marchese <56140267+cadenmarchese@users.noreply.github.com>
